PIN-5742: Deleted dialogDelegations + Added checkbox props in DialogBasic (#1205)
